+++ custard/utils.py
@@ -1,10 +1,10 @@
-from django.utils import importlib
-
+from __future__ import unicode_literals
+from importlib import import_module
 
 #==============================================================================
 def import_class(name):
     components = name.split('.')
-    mod = importlib.import_module(components[0])
+    mod = import_module(components[0])
     for comp in components[1:]:
         mod = getattr(mod, comp)
     return mod