+++ example/example/settings.py
@@ -2,10 +2,10 @@
 Django settings for example project.
 
 For more information on this file, see
-https://docs.djangoproject.com/en/1.6/topics/settings/
+https://docs.djangoproject.com/en/1.7/topics/settings/
 
 For the full list of settings and their values, see
-https://docs.djangoproject.com/en/1.6/ref/settings/
+https://docs.djangoproject.com/en/1.7/ref/settings/
 """
 
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
@@ -105,6 +105,6 @@ except ImportError:
 # Custard configuration
 
 CUSTOM_CONTENT_TYPES = (
-    'example',
-    'user',
+    'demo.example',
+    'auth.user',
 )